# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Davilla Rondeur e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /**
     * e.g. Nouveauté, Best-seller
     * @nullable
     */
  label?: string | null;
  inStock: boolean;
  featured?: boolean;
  sizes?: string[];
  colors?: string[];
  slug?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface UpdateCartItemInput {
  /** @minimum 1 */
  quantity: number;
}

export interface CreateCheckoutSessionInput {
  sessionId: string;
  email: string;
}

export interface CheckoutSessionResponse {
  url: string;
  orderId: number;
}

export interface ShippingAddress {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  line1?: string | null;
  /** @nullable */
  line2?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  email: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shippingAmount: number;
  total: number;
  shippingAddress?: ShippingAddress | null;
  /** @nullable */
  stripeSessionId?: string | null;
  createdAt: string;
  /** @nullable */
  paidAt?: string | null;
}

export interface NewsletterInput {
  email: string;
}

export interface SuccessResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminAuthResponse {
  token: string;
  email: string;
}

export interface AdminCreateProductInput {
  name: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  categoryId: number;
  /** @nullable */
  label?: string | null;
  inStock?: boolean;
  featured?: boolean;
  sizes?: string[];
  colors?: string[];
}

export interface AdminUpdateProductInput {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  categoryId?: number;
  /** @nullable */
  label?: string | null;
  inStock?: boolean;
  featured?: boolean;
  sizes?: string[];
  colors?: string[];
}

export interface PresignUploadInput {
  filename: string;
  contentType: string;
}

export interface PresignUploadResponse {
  uploadUrl: string;
  publicUrl: string;
  key: string;
}

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
featured?: boolean | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type GetCartParams = {
sessionId: string;
};